Rename of custom metric
