+++ library/src/test/java/com/datatorrent/lib/helper/OperatorContextTestHelper.java
@@ -64,7 +64,7 @@ public class OperatorContextTestHelper
     }
 
     @Override
-    public void sendCustomMetrics(Collection<String> metricNames)
+    public void sendMetrics(Collection<String> metricNames)
     {
       throw new UnsupportedOperationException("not supported");
     }
@@ -109,7 +109,7 @@ public class OperatorContextTestHelper
     }
 
     @Override
-    public void sendCustomMetrics(Collection<String> metricNames)
+    public void sendMetrics(Collection<String> metricNames)
     {
       throw new UnsupportedOperationException("Not supported yet.");
     }
